Stop bitching about imagemagick's weird quality knobs and replicate their bizarre behavior instead

+++ src/plan.rs
@@ -150,5 +150,5 @@ pub struct FilePlan {
 
 #[derive(Debug, Default)]
 pub struct Modifiers {
-    pub quality: Option<u8>,
+    pub quality: Option<f64>,
 }
